+++ server/src/routes/groupRoute.route.js
@@ -4,6 +4,7 @@ import {
   getGroup,
   sendGroupMessage,
   getGroupMessage,
+  deleteGroupMessage,
 } from "../controllers/group.controller.js";
 import { authenticate } from "../middleware/auth.middleware.js";
 
@@ -16,5 +17,7 @@ route.post("/message/send/:id", authenticate, sendGroupMessage);
 route.get("/message/getMessages/:id", authenticate, getGroupMessage);
 
 route.get("/getGroups/:id", authenticate, getGroup);
+
+route.delete("/message/delete/:id", authenticate, deleteGroupMessage);
 
 export default route;